initial wire up of campus styles. #63
