+++ services/ucd-lib-client/client/public/elements/app-footer.html
@@ -7,7 +7,7 @@
     max-width: var(--max-width);
 
     color: white;
-    background: var(--secondary-background-color);
+    background: var(--default-primary-color);
   }
 
   footer {
