+++ services/ucd-lib-client/client/public/elements/fin-app.html
@@ -1,4 +1,3 @@
-<style include="campus-brand"></style>
 <style include="shared-styles">
   :host {
     display: block;
@@ -139,9 +138,6 @@
       <app-record id="record"></app-record>
       <app-about id="about"></app-about>
     </iron-pages>
-    <div>
-      <h1 class="color-light">Hello World!</h1>
-    </div>
     <div class="footer">
       <app-footer></app-footer>
     </div>
