--- services/ucd-lib-client/client/public/elements/styles/campus-brand-properties.js
@@ -1,5 +1,0 @@
-import css from "./campus-brand-properties.css"
-
-let styleEle = document.createElement('style');
-styleEle.innerHTML = css;
-document.head.appendChild(styleEle);
