+++ services/ucd-lib-client/client/public/elements/styles/campus-theme-properties.css
@@ -1,3 +1,6 @@
+/* 
+ * CUSTOM CSS PROPERTIES FOR THE UC DAVIS CAMPUS THEME
+*/
 html {
   /* PRIMARY COLORS */
   --color-aggie-blue        : #002851;
