+++ services/ucd-lib-client/client/public/elements/styles/campus-utils-spacing.css
@@ -1,0 +1,9 @@
+/*
+ * maybe we put utility CSS classes in their own files and then import as needed?
+*/
+.m-lg {
+  margin: var(--spacing-lg);
+}
+.m-md {
+  margin: var(--spacing-md);
+}
