+++ services/ucd-lib-client/client/public/elements/styles/dams-styles-properties.css
@@ -1,0 +1,59 @@
+html {
+    --default-primary-color : var(--color-aggie-blue) !important;
+    --light-primary-color   : var(--color-aggie-blue-80) !important;
+    
+    --default-secondary-color : var(--color-aggie-gold) !important;
+    --light-secondary-color   : var(--color-rain ,#9be7ff) !important;
+    --dark-secondary-color    : var(--color-gunrock , #2286c3) !important;
+
+    --medium-background-color : #B2BDCF !important;
+    --light-background-color : #D6DCE6 !important;
+    --super-light-background-color: #f3f3f3 !important;
+  
+    --text-primary-color      : var(--color-html) !important;
+    --primary-text-color      : var(--text-primary-color) !important;
+    --secondary-text-color    : var(--default-secondary-color) !important;
+    --inverse-text-color      : var(--color-white) !important;
+    --gray-text               : var(--color-black-50) !important;
+    --text-disabled           : var(--gray-text) !important;
+    
+    --color-grey              : #808080;
+    --color-light-yellow      : #ede2c0;
+
+    --max-width               : 1200px;
+    --max-text-width          : 650px;
+
+    --layout-sm               : 768px;
+    --grid-cell-width         : 250px;
+
+    --default-outline         : 2px dotted var(--default-secondary-color);
+    
+    --paper-icon-button-ink-color : transparent;
+
+   /**
+    * Custom mixins
+    */
+    --fin-search-box-select: {
+      font-size: var(--fs-p);
+      font-weight: var(--fw-regular);
+      color: var(--default-primary-color);
+      background-color: var(--default-secondary-color);
+      width: 100px;
+    }
+    --fin-search-box-select-inverse: {
+      @apply(--fin-search-box-select);
+      color: var(--default-secondary-color);
+      background-color: var(--default-primary-color);
+    }
+    --fin-search-box-input: {
+      padding-left: 10px;
+      font-size: var(--fs-p);
+      background-color: white;
+    }
+    --cork-drop-down-arrow-color: var(--default-primary-color);
+    --cork-toggle-panel-label : {
+      padding: 10px 0;
+      color: var(--default-primary-color);
+      font-weight: var(--fw-regular-bold);
+    }
+  }
