+++ services/ucd-lib-client/client/public/elements/styles/dams-styles.css
@@ -1,0 +1,100 @@
+/* Default typography */
+paper-tabs {
+  --paper-tabs-selection-bar: {
+    border-bottom: 3px solid var(--default-primary-color);
+  }
+}
+paper-tab {
+  --paper-tab-content-unselected : {    
+    text-transform: uppercase;
+    color: var(--text-disabled);
+  }
+  --paper-tab-content : {
+    text-transform: uppercase;
+    color: var(--default-primary-color);
+  }
+}
+
+button {
+  margin: 0;
+  padding: 0;
+}
+
+input {
+  border-radius: 0;
+}
+
+input, select, button {
+  font-size        : var(--fs-p);
+  font-weight      : var(--fw-regular);
+  color            : var(--text-primary-color);
+  font-family      : proxima-nova,"Lucida Grande","Lucida Sans","Helvetica Neue",Helvetica,Arial,sans-serif;
+  }
+
+a:focus, button:focus, select:focus, div[tabindex]:focus {
+  outline: var(--default-outline);
+}
+
+paper-icon-button {
+  border: 2px solid transparent;
+  /*background-color: var(--default-primary-color);*/
+}
+
+paper-icon-button:focus {
+  border: var(--default-outline);
+  border-radius: 20px;
+}
+
+main {
+  display: flex;
+  justify-content: center;
+}
+
+main > * {
+  max-width: 1000px;
+  width: 100%;
+}
+
+ul.menu {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+
+a.italic {
+  color: var(--default-secondary-color);
+  font-style: italic;
+}
+
+a.gold {
+  color: var(--default-secondary-color);
+}
+
+.text-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.text-container > * {
+  max-width: var(--max-text-width);
+  width: 100%;
+}
+
+.container {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
+
+.container > * {
+  max-width: var(--max-width);
+  width: 100%;
+  justify-content: center;
+}
+
+fin-search-box > iron-icon {
+  color: var(--default-secondary-color);
+}
