+++ services/ucd-lib-client/client/public/elements/styles/shared-styles.js
@@ -1,5 +1,18 @@
-import styles from "./shared-styles.html"
+import ucdCssProps from "./campus-theme-properties.css";
+import ucdCss from "./campus-theme.css";
+import damsCssProps from "./dams-styles-properties.css";
+import damsCss from "./dams-styles.css";
 
-const sharedStyles = document.createElement('template');
-sharedStyles.innerHTML = styles;
-document.head.appendChild(sharedStyles.content);
+// Insert campus and DAMS style properties
+let styleEle = document.createElement('style');
+styleEle.innerHTML = ucdCssProps + damsCssProps;
+document.head.appendChild(styleEle);
+
+// Insert campus and DAMS style rules
+let templateEle = document.createElement('template');
+templateEle.innerHTML = `<style>${ucdCss + damsCss}</style>`;
+let dmEle = document.createElement('dom-module');
+dmEle.id = 'shared-styles';
+dmEle.appendChild(templateEle);
+document.head.appendChild(dmEle);
+
