--- services/ucd-lib-client/client/public/elements/styles/style-properties.html
@@ -1,66 +1,0 @@
-<custom-style>
-  <style>
-    html {
-      /**
-      * All of the colorz
-      */
-      --default-primary-color : #022851;
-      --light-primary-color   : #335379;
-      
-      --default-secondary-color : #daaa00;
-      --light-secondary-color   : #9be7ff;
-      --dark-secondary-color    : #2286c3;
-
-      --medium-background-color : #B2BDCF;
-      --light-background-color : #D6DCE6;
-      --super-light-background-color: #f3f3f3;
-    
-      --text-primary-color      : black;
-      --primary-text-color      : var(--text-primary-color);
-      --secondary-text-color    : var(--default-secondary-color);
-      --inverse-text-color      : white;
-      --gray-text               : #8B8B8B;
-      --text-disabled           : var(--gray-text);
-      
-      --color-grey              : #808080;
-      --color-light-yellow      : #ede2c0;
-
-      --max-width               : 1200px;
-      --max-text-width          : 650px;
-
-      --layout-sm               : 768px;
-      --grid-cell-width         : 250px;
-
-      --default-outline         : 2px dotted var(--default-secondary-color);
-      
-      --paper-icon-button-ink-color : transparent;
-
-     /**
-      * Custom mixins
-      */
-      --fin-search-box-select: {
-        font-size: var(--fs-p);
-        font-weight: var(--fw-regular);
-        color: var(--default-primary-color);
-        background-color: var(--default-secondary-color);
-        width: 100px;
-      }
-      --fin-search-box-select-inverse: {
-        @apply(--fin-search-box-select);
-        color: var(--default-secondary-color);
-        background-color: var(--default-primary-color);
-      }
-      --fin-search-box-input: {
-        padding-left: 10px;
-        font-size: var(--fs-p);
-        background-color: white;
-      }
-      --cork-drop-down-arrow-color: var(--default-primary-color);
-      --cork-toggle-panel-label : {
-        padding: 10px 0;
-        color: var(--default-primary-color);
-        font-weight: var(--fw-regular-bold);
-      }
-    }
-  </style>
-</custom-style>
