--- services/ucd-lib-client/client/public/elements/styles/style-properties.js
@@ -1,7 +1,0 @@
-import "@polymer/paper-styles/default-theme"
-
-import styles from "./style-properties.html"
-let styleWrapper = document.createElement('div');
-styleWrapper.style.display = 'none';
-styleWrapper.innerHTML = styles;
-document.head.appendChild(styleWrapper);
