--- services/ucd-lib-client/client/public/elements/styles/theme-properties.css
@@ -1,3 +1,0 @@
-html {
-  --testing-color: #00f10c;
-}
