--- services/ucd-lib-client/client/public/elements/styles/theme-properties.js
@@ -1,5 +1,0 @@
-import css from "./theme-properties.css"
-
-let styleEle = document.createElement('style');
-styleEle.innerHTML = css;
-document.head.appendChild(styleEle);
