--- services/ucd-lib-client/client/public/elements/styles/theme.css
@@ -1,4 +1,0 @@
-div.test {
-  background-color : var(--testing-color);
-  height: 500px;
-}
